requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-mysql==3.2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-simple-captcha==0.5.11
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Faker==1.0.7
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.1
MarkupSafe==1.1.1
Pillow==6.0.0
psycopg2==2.8.1
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
xlwt==1.3.0
